simplify breakout manual ingest flow
