+++ .github/ACDbot/BREAKOUT_WORKFLOW.md
@@ -51,24 +51,28 @@ Upload the recording. Any title is fine — the pipeline does not parse it. `unl
 
 Grab the final URL (e.g. `https://www.youtube.com/watch?v=abc123`).
 
-## Step 2: Run `init` locally
+## Step 2: Run `init` from the breakout export folder
 
 Example for the ePBS breakout from `acdt/077`:
 
 ```bash
-cd /Users/lucy/fun/pm/.github/ACDbot
+cd "/Users/lucy/Downloads/drive-download-.../2026-04-13 ... -- EPBS BREAKOUT"
 
-uv run scripts/import_manual_call.py init \
-  --series epbs \
+ACDBOT=/path/to/pm/.github/ACDbot
+
+uv run --project "$ACDBOT" \
+  "$ACDBOT/scripts/import_manual_call.py" init \
   --parent acdt/077 \
-  --source-dir "/Users/lucy/Downloads/drive-download-.../2026-04-13 ... -- EPBS BREAKOUT" \
   --youtube-url "https://www.youtube.com/watch?v=REPLACE_ME"
 ```
 
+If the folder name does not clearly contain `epbs`, `bal`, or `focil`, add `--series` explicitly.
+
 What `init` does:
 - resolves the parent call from PM metadata
+- infers the breakout series from the current folder name when possible
 - assigns the next breakout number automatically
-- normalizes the local `chat.txt` into PM's canonical chat format
+- reads and normalizes the local `chat.txt` from the current directory
 - writes `config.json` (including `videoUrl`, `parent`)
 - regenerates `artifacts/manifest.json`
 
@@ -77,7 +81,7 @@ No transcription, no API calls. Runs in seconds.
 ## Step 3: Commit and push PM changes
 
 ```bash
-cd /Users/lucy/fun/pm
+cd /path/to/pm
 
 git add .github/ACDbot/artifacts
 git commit -m "init <series> breakout for <parent>"
@@ -89,7 +93,7 @@ git push fork HEAD
 Once the PM changes land on `master`:
 
 1. `dispatch-forkcast-on-assets.yml` fires → Forkcast syncs the call with chat + video (no transcript yet).
-2. `breakout-transcription.yml` fires in parallel → downloads audio via `yt-dlp`, transcribes, runs changelog + summary, commits the derived artifacts to `master`, then dispatches Forkcast again.
+2. `breakout-transcription.yml` fires in parallel → downloads audio via `yt-dlp`, transcribes, runs changelog + summary, commits the derived artifacts to `master`, and explicitly dispatches Forkcast.
 3. Forkcast re-syncs and picks up the transcript + summary.
 
 Both CI runs are automatic after merge. If the YouTube upload is still processing when CI runs and `yt-dlp` fails, just re-run the `Breakout Transcription` workflow from the Actions tab.
